:rocket: API de Usuários funcionando para testes e alteração nas dependências do projeto no pom.xml.

+++ back-end/src/main/java/com/mateuss/financeiro_api/ImplementacaoService/UsuarioImplService.java
@@ -3,13 +3,13 @@ package com.mateuss.financeiro_api.ImplementacaoService;
 
 import com.mateuss.financeiro_api.dto.UsuarioDTORequest;
 import com.mateuss.financeiro_api.dto.UsuarioDTOResponse;
+import com.mateuss.financeiro_api.exceptions.ResourceAlreadyExistsException;
+import com.mateuss.financeiro_api.exceptions.ResourceNotFoundException;
 import com.mateuss.financeiro_api.model.Usuario;
 import com.mateuss.financeiro_api.repository.UsuarioRepository;
 import com.mateuss.financeiro_api.service.UsuarioService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Service
 public class UsuarioImplService implements UsuarioService {
@@ -17,9 +17,11 @@ public class UsuarioImplService implements UsuarioService {
     @Autowired
     UsuarioRepository usuarioRepository;
 
-
     @Override
     public void adicionarUsuario(UsuarioDTORequest usuarioDTORequest) {
+        if (usuarioRepository.existsByEmail(usuarioDTORequest.getEmail())) {
+            throw new ResourceAlreadyExistsException("Já existe um usuário com esse e-mail.");
+        }
         Usuario novoUsuario = new Usuario();
         novoUsuario.setEmail(usuarioDTORequest.getEmail());
         novoUsuario.setNome(usuarioDTORequest.getNome());
@@ -30,40 +32,37 @@ public class UsuarioImplService implements UsuarioService {
 
     @Override
     public UsuarioDTOResponse lerUsuario(Long id) {
-        Optional<Usuario> usuario = usuarioRepository.findById(id);
-        if (usuario.isPresent()){
-            UsuarioDTOResponse usuarioDTOResponse = new UsuarioDTOResponse();
-            usuarioDTOResponse.setEmail(usuario.get().getEmail());
-            usuarioDTOResponse.setNome(usuario.get().getNome());
-            usuarioDTOResponse.setSobrenome(usuario.get().getSobrenome());
-            return usuarioDTOResponse;
-        }
-        return null;
+        Usuario usuario = usuarioRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Usuário com ID " + id + " não encontrado."));
+        UsuarioDTOResponse usuarioDTOResponse = new UsuarioDTOResponse();
+        usuarioDTOResponse.setEmail(usuario.getEmail());
+        usuarioDTOResponse.setNome(usuario.getNome());
+        usuarioDTOResponse.setSobrenome(usuario.getSobrenome());
+        return usuarioDTOResponse;
+    }
+
+    public UsuarioDTOResponse atualizarUsuario(UsuarioDTORequest usuarioDTORequest, Long id) {
+        Usuario usuarioExistente = usuarioRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Usuário com ID " + id + " não encontrado para atualização de dados."));
+
+        UsuarioDTOResponse usuarioDTOResponse = new UsuarioDTOResponse();
+        usuarioExistente.setEmail(usuarioDTORequest.getEmail());
+        usuarioExistente.setNome(usuarioDTORequest.getNome());
+        usuarioExistente.setSobrenome(usuarioDTORequest.getSobrenome());
+        usuarioExistente.setSenha(usuarioDTORequest.getSenha());
+        usuarioRepository.save(usuarioExistente);
+
+        usuarioDTOResponse.setNome(usuarioDTORequest.getNome());
+        usuarioDTOResponse.setSobrenome(usuarioDTORequest.getSobrenome());
+        usuarioDTOResponse.setEmail(usuarioDTORequest.getEmail());
+        return usuarioDTOResponse;
+
     }
 
     @Override
-    public UsuarioDTOResponse atualizarUsuario(UsuarioDTORequest usuarioDTORequest) {
-        Usuario usuarioExistente = usuarioRepository.findByEmail(usuarioDTORequest.getEmail());
-        if (usuarioExistente != null) {
-            UsuarioDTOResponse usuario = new UsuarioDTOResponse();
-            usuarioExistente.setEmail(usuarioDTORequest.getEmail());
-            usuarioExistente.setNome(usuarioDTORequest.getNome());
-            usuarioExistente.setSobrenome(usuarioDTORequest.getSobrenome());
-            usuarioExistente.setSenha(usuarioDTORequest.getSenha());
-            usuarioRepository.save(usuarioExistente);
-
-            usuario.setSobrenome(usuarioDTORequest.getSobrenome());
-            usuario.setNome(usuarioDTORequest.getNome());
-            usuario.setEmail(usuarioDTORequest.getEmail());
-            return usuario;
+    public void deletarUsuario(Long id) {
+        if (usuarioRepository.existsById(id)) {
+            usuarioRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("Usuário com ID " + id + " não encontrado para exclusão.");
         }
-        return null;
-    }
-
-    @Override
-    public UsuarioDTOResponse deletarUsuario(Long id) {
-        UsuarioDTOResponse usuario = lerUsuario(id);
-        usuarioRepository.deleteById(id);
-        return usuario;
     }
 }
